netflix_recs/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1

pandas~=2.2
pyarrow
openpyxl
plotly
scikit-learn
